+++ services/watchlist_services.py
@@ -6,12 +6,12 @@ from flask import jsonify
 
 
 
-def add_watch_list_item(userID,Media_TMDB_ID,Parent_ID):
+def add_watch_list_item(userID,Media_TMDB_ID,Parent_ID, is_movie):
     try:
-        insert_query = f"INSERT INTO `final_project_db`.`watch_lists_objects`(`User_ID`,`TMDB_ID`,`Parent_ID`) VALUES (%s,%s,%s) "
-        cursor.execute(insert_query, (userID,Media_TMDB_ID ,Parent_ID))
+        insert_query = f"INSERT INTO `final_project_db`.`watch_lists_objects`(`User_ID`,`TMDB_ID`,`Parent_ID`, `is_movie`) VALUES (%s,%s,%s, %s) "
+        cursor.execute(insert_query, (userID,Media_TMDB_ID ,Parent_ID, is_movie))
         connection.commit()
-        return jsonify({"type":"success" ,"massage":"Complited"})
+        return "Added movie {} to watchlist {}".format(Media_TMDB_ID, Parent_ID)
     except mysql.connector.Error as err:
         if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
             print("Something is wrong with your user name or password")
@@ -30,7 +30,11 @@ def remove_watch_list_item(userID,watch_list_item_id):
         delete_query = "DELETE FROM `final_project_db`.`watch_lists_objects` WHERE `ID`= %s AND `User_ID`= %s ;"
         cursor.execute(delete_query, (watch_list_item_id, userID))
         connection.commit()
-        return jsonify({"type":"success" ,"massage":"Complited"})
+        if cursor.rowcount > 0:
+            print(f"Deletion of watchlist item {watch_list_item_id} successful")
+            return jsonify({"type": "success", "message": f"Removed item {watch_list_item_id}"}), 200
+        else:
+            return jsonify({"type": "Failure", "message": "No item deleted. Please make sure it belongs to the logged in user"}), 200
     except mysql.connector.Error as err:
         if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
             print("Something is wrong with your user name or password")
@@ -44,25 +48,36 @@ def remove_watch_list_item(userID,watch_list_item_id):
 
 
 def create_watchlist(user_id,name,Is_main):
+    # Query to get the last inserted id for the given user_id
+    watchlist_id_query = """
+        SELECT id 
+        FROM `final_project_db`.`watch_lists_names` 
+        WHERE User_ID = %s 
+        ORDER BY id DESC 
+        LIMIT 1
+    """
     try:
      if (Is_main):
          query = f"SELECT EXISTS(SELECT 1 FROM `final_project_db`.`users` WHERE User_ID = %s AND Main= %s)"
          cursor.execute(query, (user_id,True))
          exists = cursor.fetchone()[0]
-
          if not exists:
              # Insert the ID if it does not exist
              insert_query = f"INSERT INTO `final_project_db`.`watch_lists_names` User_ID,name,Main) VALUES (%s,%s,%s)"
              cursor.execute(insert_query, (user_id,"Main",True ))
              connection.commit()
              print(f"ID {user_id}  list was added to the table .")
-             return jsonify({"ID {user_id}  list was added to the table ."}), 404
+             cursor.execute(watchlist_id_query, user_id)
+             return cursor.fetchone()[0]
          else:
-             print(f"ID {user_id}  list mainalready exists in the table .")
+             print(f"ID {user_id}  list main watchlist already exists in the table .")
      else :
          insert_query = f"INSERT INTO `final_project_db`.`watch_lists_names` (User_ID,name,Main) VALUES (%s,%s,%s)"
          cursor.execute(insert_query, (user_id, name, False))
          connection.commit()
+         cursor.execute(watchlist_id_query, (user_id,))
+         new_id = cursor.fetchone()[0]
+         return new_id
     except mysql.connector.Error as err:
 
         if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
@@ -81,13 +96,20 @@ def create_watchlist(user_id,name,Is_main):
 
 def remove_watch_list(userID,watch_list_id):
     try:
+        # Delete watchlist object, not items
         delete_query1 = "DELETE FROM `final_project_db`.`watch_lists_names` WHERE `ID`= %s AND `User_ID`= %s ;"
-
         cursor.execute(delete_query1, (watch_list_id, userID))
-        delete_query2 = "DELETE FROM `final_project_db`.`watch_lists_objects` WHERE `Parent_ID`= %s AND `User_ID`= %s ;"
-        cursor.execute(delete_query2, (watch_list_id, userID))
         connection.commit()
-        return jsonify({"type":"success" ,"massage":"Complited"})
+        if cursor.rowcount > 0:
+            # Delete items from watchlist
+            delete_query2 = "DELETE FROM `final_project_db`.`watch_lists_objects` WHERE `Parent_ID`= %s AND `User_ID`= %s ;"
+            cursor.execute(delete_query2, (watch_list_id, userID))
+            connection.commit()
+            print(f"Deletion of watchlist {watch_list_id} successful")
+            return jsonify({"type": "success", "message": f"Removed watchlist {watch_list_id}"}), 200
+        else:
+            return jsonify({"type": "Failure", "message": "No watchlist deleted. Please make sure it belongs to the logged in user"}), 200
+        # Delete items from watchlist
     except mysql.connector.Error as err:
         if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
             print("Something is wrong with your user name or password")
@@ -100,14 +122,10 @@ def remove_watch_list(userID,watch_list_id):
             return err
 			
 def get_user_watchlists(user_id):
-
     try:
-
          query = f"SELECT * FROM `final_project_db`.`watch_lists_names` WHERE User_ID = %s"
          cursor2.execute(query, (user_id,))
-         return jsonify(cursor2.fetchall())
-
-
+         return cursor2.fetchall()
 
     except mysql.connector.Error as err:
         if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
@@ -128,6 +146,26 @@ def get_watchlist_by_id(watchlist_ID):
         cursor2.execute(query, (watchlist_ID,))
         results = cursor2.fetchall()
         return results
+
+    except mysql.connector.Error as err:
+        if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
+            print("Something is wrong with your user name or password")
+            return jsonify({"type":"Error" ,"message":"Something is wrong with your user name or password"}), 404
+        elif err.errno == errorcode.ER_BAD_DB_ERROR:
+            print("Database does not exist")
+            return jsonify({"type":"Error" ,"message":"Database does not exist"}), 404
+        else:
+            print(err)
+            return err
+
+
+def get_watchlist_details_only(watchlist_ID):
+    try:
+
+        query = f"SELECT * FROM `final_project_db`.`watch_lists_names` WHERE ID = %s"
+        cursor2.execute(query, (watchlist_ID,))
+        results = cursor2.fetchall()
+        return results[0]
 
     except mysql.connector.Error as err:
         if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
